--- src/main/java/calculator/Numbers.java
@@ -1,11 +1,0 @@
-package calculator;
-
-import java.util.List;
-
-public class Numbers {
-    private List<Integer> numbers;
-
-    public Numbers(List<Integer> numbers) {
-        this.numbers = numbers;
-    }
-}
