+++ src/main/java/calculator/constant/MessageConstant.java
@@ -1,6 +1,5 @@
-package calculator;
+package calculator.constant;
 
 public class MessageConstant {
     public static final String NEGATIVE_NUMBER_EXCEPTION = "음수는 포함될 수 없습니다.";
-
 }
